// block-performance-1.0.0/src/library/url/search-params-merge/type.ts
// auto generated by dev/system

import { Logger } from "../../../myutils/logger.js";

export interface Input {
    url: string;
    search_params?: { [key: string]: string };
}

export interface OutputOk {
    url: string;
}

export interface OutputFail {}

export interface Callback<R> {
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}

export function check_input<R>(plog: Logger, v: any, cb: { ok: () => R; fail: (err: Error) => R }): R {
    const log = plog.sub("check_input");
    log.variable("v", v);
    try {
        log.println("v must be object");
        if (typeof v !== "object" || v === null) {
            throw new Error("v is not object");
        }

        Object.keys(v).forEach((field) => {
            if (field === "url") return;
            if (field === "search_params") return;
            throw new Error("v contains unknown field: " + field);
        });

        log.println("v.url must be string");
        if (typeof v.url !== "string") {
            throw new Error("v.url is not string");
        }

        if (v.search_params !== undefined) {
            log.println("v.search_params must be object");
            if (typeof v.search_params !== "object" || v.search_params === null) {
                throw new Error("v.search_params is not object");
            }

            Object.keys(v.search_params).forEach((field) => {
                // a dynamic field, check it (FIXME log message is not clear)

                log.println("v.search_params[field] must be string");
                if (typeof v.search_params[field] !== "string") {
                    throw new Error("v.search_params[field] is not string");
                }
            });
        }
    } catch (err) {
        log.error(err);
        return cb.fail(err);
    }
    // nothing wrong
    log.ok();
    return cb.ok();
}

// JSON stringify value before copy to handle some specific problem
// eg. Date Object probelm
export function copy_output_ok(v: any): OutputOk {
    if (v !== undefined) {
        const v_cloned = JSON.parse(JSON.stringify(v));
        return _copy_output_ok(v_cloned);
    } else {
        return _copy_output_ok(v);
    }
}

export function _copy_output_ok(v: any): OutputOk {
    if (typeof v === "object" && v !== null) {
        const obj = {
            url: copy_url(v.url)
        };
        return obj;
    } else {
        return { url: "" };
    }

    function copy_url(v: any): string {
        return typeof v === "string" ? v : "";
    }
}

// backward compatible purpose
export type Output = OutputOk;
export const copy_output = copy_output_ok;